+++ CHANGELOG.md
@@ -40,7 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial key functionality
 
-[Unreleased]: https://github.com/gyselroth/kube-ldap/compare/master...dev
+[Unreleased]: https://github.com/gyselroth/kube-ldap/compare/v1.3.0...master
 [1.3.0]: https://github.com/gyselroth/kube-ldap/compare/v1.2.1...v1.3.0
 [1.2.1]: https://github.com/gyselroth/kube-ldap/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/gyselroth/kube-ldap/compare/v1.1.0...v1.2.0